refactor: concurrent docs 파일 6개 → ConcurrentDocs 단일 파일로 통합
